requirements.txt
flask
pyyaml
opencv-python